third_party/blink/web_tests/external/wpt/css/css-masking/parsing/clip-path-shape-parsing-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['clip-path'] = "shape(from 10px 10px, curve to 50px 20px with 10rem 1%)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['clip-path'] = "shape(from 10px 10px, curve to 50px 20px with 10rem 1px / 20vh 1ch)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['clip-path'] = "shape(from 10px 10px, curve by 50px 20px with 10rem 1px / 20vh 1ch)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['clip-path'] = "shape(from 10px 10px, smooth to 50px 20px with 10rem 1%)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['clip-path'] = "shape(from 10px 10px, smooth to 50px 1pt)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['clip-path'] = "shape(from 10% 1rem, arc to 50px 1pt of 20% cw large rotate 25deg)" should set the property value
  assert_equals: serialization should be canonical expected "shape(from 10% 1rem, arc to 50px 1pt of 20% cw large rotate 25deg)" but got "shape(from 10% 1rem, arc to 50px 1pt of 20% rotate 25deg)"
Harness: the test ran to completion.
